backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "portfolio-backend"
version = "1.0.0"
description = "Portfolio backend with Django REST Framework"
requires-python = ">=3.11"
dependencies = [
    "Django==5.0.0",
    "djangorestframework==3.14.0",
    "djangorestframework-simplejwt==5.3.2",
    "django-cors-headers==4.3.1",
    "psycopg==3.1.15",
    "psycopg-binary==3.1.15",
    "python-decouple==3.8",
    "whitenoise==6.6.0",
    "gunicorn==21.2.0",
    "pillow==10.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-django==4.7.0",
    "pytest-cov==4.1.0",
    "black==23.12.0",
    "flake8==6.1.0",
    "isort==5.13.2",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100
